+++ packages/collector/test/tracing/common/test.js
@@ -7,8 +7,8 @@ const supportedVersion = require('@instana/core').tracing.supportedVersion;
 const config = require('../../../../core/test/config');
 const delay = require('../../../../core/test/test_util/delay');
 const testUtils = require('../../../../core/test/test_util');
+const ProcessControls = require('../ProcessControls');
 
-let Controls;
 const extendedTimeout = Math.max(config.getTestTimeout(), 10000);
 
 describe('tracing/common', function() {
@@ -16,13 +16,13 @@ describe('tracing/common', function() {
     return;
   }
 
-  Controls = require('./controls');
   describe('delay', function() {
     describe('with minimal delay', function() {
       this.timeout(extendedTimeout);
       const agentControls = setupAgentControls();
-      const controls = new Controls({
+      const controls = new ProcessControls({
         agentControls,
+        dirname: __dirname,
         minimalDelay: 6000
       });
       registerDelayTest.call(this, agentControls, controls, true);
@@ -31,7 +31,10 @@ describe('tracing/common', function() {
     describe('without minimal delay', function() {
       this.timeout(config.getTestTimeout());
       const agentControls = setupAgentControls();
-      const controls = new Controls({ agentControls });
+      const controls = new ProcessControls({
+        agentControls,
+        dirname: __dirname
+      });
       registerDelayTest.call(this, agentControls, controls, false);
     });
 
@@ -76,7 +79,7 @@ describe('tracing/common', function() {
                   expect(span.data.http.method).to.equal('GET');
                   expect(span.data.http.url).to.equal('/');
                   expect(span.data.http.status).to.equal(200);
-                  expect(span.data.http.host).to.equal('127.0.0.1:3215');
+                  expect(span.data.http.host).to.equal('127.0.0.1:3216');
                 }),
               Math.max(extendedTimeout / 2, 10000)
             )
@@ -101,7 +104,7 @@ describe('tracing/common', function() {
           expect(span.data.http.method).to.equal('GET');
           expect(span.data.http.url).to.equal('/');
           expect(span.data.http.status).to.equal(200);
-          expect(span.data.http.host).to.equal('127.0.0.1:3215');
+          expect(span.data.http.host).to.equal('127.0.0.1:3216');
         })
       );
     }
@@ -110,8 +113,9 @@ describe('tracing/common', function() {
   describe('service name', function() {
     describe('with env var', function() {
       const agentControls = setupAgentControls();
-      const controls = new Controls({
+      const controls = new ProcessControls({
         agentControls,
+        dirname: __dirname,
         env: {
           INSTANA_SERVICE_NAME: 'much-custom-very-wow service'
         }
@@ -121,8 +125,9 @@ describe('tracing/common', function() {
 
     describe('with config', function() {
       const agentControls = setupAgentControls();
-      const controls = new Controls({
+      const controls = new ProcessControls({
         agentControls,
+        dirname: __dirname,
         env: {
           // this makes the app set the serviceName per config object
           SERVICE_CONFIG: 'much-custom-very-wow service'
@@ -133,13 +138,19 @@ describe('tracing/common', function() {
 
     describe('with header when agent is configured to capture the header', function() {
       const agentControls = setupAgentControls(true);
-      const controls = new Controls({ agentControls });
+      const controls = new ProcessControls({
+        agentControls,
+        dirname: __dirname
+      });
       registerServiceNameTest.call(this, agentControls, controls, 'X-Instana-Service header', true);
     });
 
     describe('with header when agent is _not_ configured to capture the header', function() {
       const agentControls = setupAgentControls(false);
-      const controls = new Controls({ agentControls });
+      const controls = new ProcessControls({
+        agentControls,
+        dirname: __dirname
+      });
       registerServiceNameTest.call(this, agentControls, controls, 'X-Instana-Service header', false);
     });
 
@@ -175,6 +186,73 @@ describe('tracing/common', function() {
         })
       );
     }
+  });
+
+  describe('disable individual tracers', function() {
+    this.timeout(config.getTestTimeout());
+
+    describe('disable an individual tracer', () => {
+      const agentControls = setupAgentControls();
+      const controls = new ProcessControls({
+        agentControls,
+        dirname: __dirname,
+        env: {
+          INSTANA_DISABLED_TRACERS: 'pino'
+        }
+      });
+      controls.registerTestHooks();
+
+      it('can disable a single instrumentation', () =>
+        controls
+          .sendRequest({
+            path: '/with-log'
+          })
+          .then(() => {
+            return testUtils.retry(() =>
+              agentControls.getSpans().then(spans => {
+                expect(spans.length).to.equal(1);
+                expect(spans[0].n).to.equal('node.http.server');
+              })
+            );
+          }));
+    });
+
+    describe('robustness against overriding Array.find', () => {
+      const agentControls = setupAgentControls();
+      const controls = new ProcessControls({
+        agentControls,
+        dirname: __dirname,
+        env: {
+          SCREW_AROUND_WITH_UP_ARRAY_FIND: 'sure why not?'
+        }
+      });
+      controls.registerTestHooks();
+
+      // Story time: There is a package out there that overrides Array.find with different behaviour that, once upon a
+      // time, as a random side effect, disabled our auto tracing. This is a regression test to make sure we are now
+      // robust against that particular breakage.
+      //
+      // It is precisely this issue that broke the activation of individual instrumentations:
+      // https://github.com/montagejs/collections/issues/178
+      //
+      // In particular, the monkey patched version of Array.find returns -1 if nothing is found, while the standard
+      // Array.find returns undefined. When checking if an array contains something with find, this reverses the logic
+      // of the check because -1 is truthy and undefined is falsy.
+
+      it('messing up Array.find must not break tracing', () =>
+        controls
+          .sendRequest({
+            path: '/'
+          })
+          .then(() => {
+            return testUtils.retry(() =>
+              agentControls.getSpans().then(spans => {
+                expect(spans.length).to.equal(1);
+                expect(spans[0].n).to.equal('node.http.server');
+              })
+            );
+          }));
+    });
   });
 
   function setupAgentControls(captureXInstanaServiceHeader) {